+++ internal/cli/termui/dot_progress.go
