+++ internal/cli/termui/spin_progress.go
